[Silver V] Title: 타일 장식물, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13301. 타일 장식물/타일 장식물.cc
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 
-int d[81];
+long long d[81];
 int main()
 {
 	int n;
